requirements.txt
Django>=3.2.8
djangorestframework>=3.12.4
psycopg2>=2.7.5,<2.8.0
django-cors-headers>=3.6.0,<3.7.0
celery>=4.4.2,<5.0
redis
django-redis
